+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to
   tests.
 
 ### Fixed
+- Support database files with both `.yml` and `.yaml` extensions (#101).
 - draw: Detect when unable to find racks to draw in infrastructure and raise
   specific exception instead of crashing with `ZeroDivisionError` (#97→#100).
 - docs: